Added two losses to play with temperature annealing

+++ prior_networks/priornet/run/train_dpn.py
@@ -140,7 +140,7 @@ def main():
         if len(train_dataset) == len(ood_dataset):
             train_dataset = data.ConcatDataset([train_dataset, ood_dataset])
         elif len(train_dataset) < len(ood_dataset):
-            ratio = int(np.round(float(len(ood_dataset)) / float(len(train_dataset))))
+            ratio = np.round(float(len(ood_dataset)) / float(len(train_dataset)))
             assert ratio.is_integer()
             dataset_list = [train_dataset, ] * (1+int(ratio))
             dataset_list.append(ood_dataset)
